+++ src/entities/core-entities/outright-league/outright-league-fixture.ts
@@ -24,6 +24,10 @@ export class OutrightLeagueFixture {
   @Type(() => Date)
   lastUpdate?: Date;
 
+  @Expose({ name: 'StartDate' })
+  @Type(() => Date)
+  startDate?: Date;
+
   @Expose({ name: 'Status' })
   status?: FixtureStatus;